feat: add source names, refactor source code, change some tiles
change TileEntity to contain a source type for sources, source contains a link of the source, the
name of the source and whether it has been modified or not.
Refactor code for displaying the source
out to TilesSource.
Modify all entities to contain the required keys and values for source.
Change
vorkath (melee) to just vorkath with new tiles from runelite.
Rename commander zilyana (no-stam) to
commander zilyana (energy-efficient) to reflect the original video more accurately.
Update [entity]
test to handle new source type.
Update snapshots.